Split formatting and decoding
