+++ disassembler/decode1.c
@@ -4,7 +4,7 @@
 
 #include "operations.h"
 #include "encodings.h"
-#include "arm64dis.h"
+#include "decode.h"
 #include "decode2.h"
 #include "pcode.h"
