+++ disassembler/pcode.c
@@ -2,7 +2,8 @@
 #include <stddef.h>
 #include <stdbool.h>
 
-#include "arm64dis.h"
+#include "decode.h"
+#include "format.h"
 #include "pcode.h"
 
 int BitCount(uint32_t x)
